+++ crates/ethereum/node/src/evm.rs
@@ -1,6 +1,7 @@
 //! Ethereum EVM support
 
 #[doc(inline)]
-pub use reth_evm_ethereum::execute::EthExecutorProvider;
+#[allow(deprecated)]
+pub use reth_evm_ethereum::EthExecutorProvider;
 #[doc(inline)]
 pub use reth_evm_ethereum::{EthEvm, EthEvmConfig};
