+++ crates/ethereum/node/src/lib.rs
@@ -17,7 +17,10 @@ use revm as _;
 pub use reth_ethereum_engine_primitives::EthEngineTypes;
 
 pub mod evm;
-pub use evm::{EthEvmConfig, EthExecutorProvider};
+pub use evm::EthEvmConfig;
+
+#[allow(deprecated)]
+pub use evm::EthExecutorProvider;
 
 pub use reth_ethereum_consensus as consensus;
 pub mod node;
