+++ crates/exex/exex/src/manager.rs
@@ -663,7 +663,7 @@ mod tests {
     use futures::{StreamExt, TryStreamExt};
     use rand::Rng;
     use reth_db_common::init::init_genesis;
-    use reth_evm_ethereum::{execute::EthExecutorProvider, EthEvmConfig};
+    use reth_evm_ethereum::EthEvmConfig;
     use reth_primitives_traits::RecoveredBlock;
     use reth_provider::{
         providers::BlockchainProvider, test_utils::create_test_provider_factory, BlockReader,
@@ -1107,7 +1107,7 @@ mod tests {
             "test_exex".to_string(),
             Default::default(),
             provider,
-            EthExecutorProvider::mainnet(),
+            EthEvmConfig::mainnet(),
             wal.handle(),
         );
 
@@ -1162,7 +1162,7 @@ mod tests {
             "test_exex".to_string(),
             Default::default(),
             provider,
-            EthExecutorProvider::mainnet(),
+            EthEvmConfig::mainnet(),
             wal.handle(),
         );
 
@@ -1212,7 +1212,7 @@ mod tests {
             "test_exex".to_string(),
             Default::default(),
             provider,
-            EthExecutorProvider::mainnet(),
+            EthEvmConfig::mainnet(),
             wal.handle(),
         );
 
@@ -1255,7 +1255,7 @@ mod tests {
             "test_exex".to_string(),
             Default::default(),
             provider,
-            EthExecutorProvider::mainnet(),
+            EthEvmConfig::mainnet(),
             wal.handle(),
         );
 
@@ -1315,7 +1315,7 @@ mod tests {
             "test_exex".to_string(),
             Default::default(),
             provider.clone(),
-            EthExecutorProvider::mainnet(),
+            EthEvmConfig::mainnet(),
             wal.handle(),
         );
